+++ pat/discoproxyclient.py
@@ -94,15 +94,16 @@ class DiscoProxyClient:
         self._my_server_port = None
         self.subscribe_response = None
 
-    def start(self, subscribe: bool = True):
+    def start(self, subscribe: bool = True, http_server_start_timeout: float = 60.0):
         """Subscribe."""
         # first start http server, the services need to know the ip port number
+        self._logger.info('Starting http server ...')
         self._http_server.start()
 
-        event_is_set = self._http_server.started_evt.wait(timeout=15.0)
-        if not event_is_set:
-            self._logger.error('Cannot start device, start event of http server not set.')
-            raise RuntimeError('Cannot start device, start event of http server not set.')
+        if not self._http_server.started_evt.wait(timeout=http_server_start_timeout):
+            msg = f'Http server could not be started within {http_server_start_timeout} seconds.'
+            raise RuntimeError(msg)
+        self._logger.info('Http server started. Serving EventSink on %s', self._http_server.base_url)
         self._my_server_port = self._http_server.my_port
         self._http_server.dispatcher.register_instance('', self._msg_converter)
 
@@ -350,8 +351,11 @@ if __name__ == '__main__':
         for the_service in the_services:
             log.info('found service: %r', the_service)
 
-    def main():
-        """Execute disco proxy."""
+    def main(http_server_start_timeout: float = 60.0):
+        """Execute disco proxy.
+
+        param http_server_start_timeout: timeout to start the internal http server
+        """
         # example code how to use the DiscoProxyClient.
         # It assumes a discovery proxy is reachable on disco_ip address.
         basic_logging_setup()
@@ -379,7 +383,7 @@ if __name__ == '__main__':
         )
 
         proxy = DiscoProxyClient(disco_ip, my_ip, ssl_contexts)
-        proxy.start()
+        proxy.start(http_server_start_timeout=http_server_start_timeout)
         try:
             services = proxy.search_services()
             log_services(logger, services)