+++ src/sdc11073/httpserver/httpserverimpl.py
@@ -1,15 +1,23 @@
+"""HTTP server implementation with threading and request dispatching."""
+
 from __future__ import annotations
 
 import logging
 import socket
 import threading
-import traceback
 from dataclasses import dataclass
 from http.server import HTTPServer
+from typing import TYPE_CHECKING
 
-from .httprequesthandler import DispatchingRequestHandler
 from sdc11073.dispatch import PathElementRegistry
 from sdc11073.loghelper import LoggerAdapter
+
+from .httprequesthandler import DispatchingRequestHandler
+
+if TYPE_CHECKING:
+    from collections.abc import Iterable
+
+    from sdc11073 import certloader
 
 
 @dataclass(frozen=True)
@@ -20,11 +28,15 @@ class _ThreadInfo:
 
 
 class _ThreadingHTTPServer(HTTPServer):
-    """ Each request is handled in a thread.
-    """
+    """Each request is handled in a thread."""
 
-    def __init__(self, logger, server_address,
-                 chunk_size, supported_encodings):
+    def __init__(
+        self,
+        logger: LoggerAdapter,
+        server_address: tuple[str, int],
+        chunk_size: int,
+        supported_encodings: Iterable[str],
+    ):
         self.daemon_threads = True
         self.threads = []
         self.logger = logger
@@ -33,14 +45,13 @@ class _ThreadingHTTPServer(HTTPServer):
         self.supported_encodings = supported_encodings
         super().__init__(server_address, DispatchingRequestHandler)
 
-    def process_request_thread(self, request, client_address):
-        """Same as in BaseServer but as a thread.
-        """
+    def process_request_thread(self, request, client_address):  # noqa: ANN001
+        """Same as in BaseServer but as a thread."""  # noqa: D401
         try:
             self.finish_request(request, client_address)
         except (ConnectionResetError, ConnectionAbortedError) as ex:
-            self.logger.info('Connection reset by {}: {}', client_address, ex)
-        except Exception:
+            self.logger.info('Connection reset by %s: %s', client_address, ex)
+        except Exception:  # noqa: BLE001
             self.handle_error(request, client_address)
         finally:
             self.shutdown_request(request)
@@ -49,11 +60,11 @@ class _ThreadingHTTPServer(HTTPServer):
                 if thread_info.request == request:
                     self.threads.remove(thread_info)
 
-    def process_request(self, request, client_address):
+    def process_request(self, request, client_address):  # noqa: ANN001
         """Start a new thread to process the request."""
-        thread = threading.Thread(target=self.process_request_thread,
-                                  args=(request, client_address),
-                                  name=f'SubscrRecv{client_address}')
+        thread = threading.Thread(
+            target=self.process_request_thread, args=(request, client_address), name=f'SubscrRecv{client_address}',
+        )
         thread.daemon = True
         self.threads.append(_ThreadInfo(thread, request, client_address))
         thread.start()
@@ -68,21 +79,29 @@ class _ThreadingHTTPServer(HTTPServer):
                 try:
                     thread_info.request.shutdown(socket.SHUT_RDWR)
                     thread_info.request.close()
-                    self.logger.info('closed socket for notifications from {}', thread_info.client_address)
+                    self.logger.info('closed socket for notifications from %s', thread_info.client_address)
                 except OSError:
                     # the connection is already closed
                     continue
-                except Exception as ex:
-                    self.logger.warn('error closing socket for notifications from {}: {}', thread_info.client_address,
-                                     ex)
+                except Exception as ex:  # noqa: BLE001
+                    self.logger.warning(
+                        'error closing socket for notifications from %s: %s', thread_info.client_address, ex,
+                    )
 
 
 class HttpServerThreadBase(threading.Thread):
+    """A Thread running a ThreadingHTTPServer."""
 
-    def __init__(self, my_ipaddress, ssl_context, supported_encodings,
-                 logger, chunk_size=0):
-        """
-        Runs a ThreadingHTTPServer in a thread, so that it can be stopped without blocking.
+    def __init__(
+        self,
+        my_ipaddress: str,
+        ssl_context: certloader.SSLContextContainer | None,
+        supported_encodings: Iterable[str],
+        logger: logging.Logger | LoggerAdapter,
+        chunk_size: int = 0,
+    ):
+        """Run a ThreadingHTTPServer in a thread, so that it can be stopped without blocking.
+
         Handling of requests happens in two stages:
         - the http server instantiates a request handler with the request
         - the request handler forwards the handling itself to a dispatcher (due to the dynamic nature of the handling).
@@ -106,20 +125,20 @@ class HttpServerThreadBase(threading.Thread):
             self.logger = logger
         self.chunk_size = chunk_size
         # create and set up the dispatcher for all incoming requests
-        self.started_evt = threading.Event()  # helps to wait until thread has initialised is variables
+        self.started_evt = threading.Event()  # helps to wait until thread has initialized is variables
         self._stop_requested = False
         self.base_url = None
 
     def run(self):
+        """Run the http server."""
         self._stop_requested = False
         try:
             myport = 0  # zero means that OS selects a free port
-            self.httpd = _ThreadingHTTPServer(self.logger,
-                                              (self._my_ipaddress, myport),
-                                              self.chunk_size,
-                                              self.supported_encodings)
+            self.httpd = _ThreadingHTTPServer(
+                self.logger, (self._my_ipaddress, myport), self.chunk_size, self.supported_encodings,
+            )
             self.my_port = self.httpd.server_port
-            self.logger.info('starting http server on {}:{}', self._my_ipaddress, self.my_port)
+            self.logger.info('starting http server on %s:%s', self._my_ipaddress, self.my_port)
             if self._ssl_context:
                 self.httpd.socket = self._ssl_context.wrap_socket(self.httpd.socket, server_side=True)
                 self.base_url = f'https://{self._my_ipaddress}:{self.my_port}/'
@@ -130,17 +149,23 @@ class HttpServerThreadBase(threading.Thread):
             self.httpd.serve_forever()
         except Exception:
             if not self._stop_requested:
-                self.logger.error(
-                    f'Unhandled Exception at thread runtime. Thread will abort! {traceback.format_exc()}')
+                self.logger.exception('Unhandled Exception at thread runtime. Thread will abort!')
             raise
         finally:
             self.logger.info('http server stopped.')
 
     @property
-    def dispatcher(self):
+    def dispatcher(self) -> PathElementRegistry:
+        """Return the dispatcher responsible for handling requests."""
+        if not self.started_evt.is_set():
+            raise RuntimeError('http server not started yet, dispatcher not available')
         return self.httpd.dispatcher
 
     def stop(self):
+        """Stop the http server."""
+        if not self.started_evt.is_set():
+            self.logger.warning('http server was not started yet - cannot be stopped')
+            return
         self._stop_requested = True
         self.httpd.shutdown()
         self.httpd.server_close()
@@ -148,5 +173,6 @@ class HttpServerThreadBase(threading.Thread):
             if thread_info.thread.is_alive():
                 thread_info.thread.join(1)
             if thread_info.thread.is_alive():
-                self.logger.warn('could not end client thread for notifications from {}', thread_info.client_address)
+                self.logger.warning('could not end client thread for notifications from %s', thread_info.client_address)
         del self.httpd.threads[:]
+        self.started_evt.clear()