+++ src/sdc11073/mdib/mdibbase.py
@@ -1,4 +1,5 @@
 """The module contains the implementation of MdibBase plus entity interface."""
+
 from __future__ import annotations
 
 import copy
@@ -27,6 +28,7 @@ if TYPE_CHECKING:
 
     from .descriptorcontainers import AbstractDescriptorContainer, AbstractOperationDescriptorContainer
     from .statecontainers import AbstractMultiStateContainer, AbstractStateContainer
+
 
 @dataclass
 class MdibVersionGroup:
@@ -91,8 +93,10 @@ class DescriptorsLookup(_MultikeyWithVersionLookup):
         self.add_index('parent_handle', multikey.IndexDefinition(lambda obj: obj.parent_handle))
         self.add_index('NODETYPE', multikey.IndexDefinition(lambda obj: obj.NODETYPE))
         self.add_index('coding', multikey.IndexDefinition(lambda obj: obj.coding))
-        self.add_index('condition_signaled',
-                       multikey.IndexDefinition(lambda obj: obj.ConditionSignaled, index_none_values=False))
+        self.add_index(
+            'condition_signaled',
+            multikey.IndexDefinition(lambda obj: obj.ConditionSignaled, index_none_values=False),
+        )
         # an index to find all alert conditions for a metric (AlertCondition is the only class that has a
         # "Source" attribute, therefore this simple approach without type testing is sufficient):
         self.add_index('source', multikey.IndexDefinition1n(lambda obj: obj.Source, index_none_values=False))
@@ -198,10 +202,8 @@ class MultiStatesLookup(_MultikeyWithVersionLookup):
     def __init__(self):
         super().__init__()
         self.add_index('descriptor_handle', multikey.IndexDefinition(lambda obj: obj.DescriptorHandle))
-        self.add_index('handle',
-                       multikey.UIndexDefinition(lambda obj: obj.Handle, index_none_values=False))
-        self.add_index('NODETYPE',
-                       multikey.IndexDefinition(lambda obj: obj.NODETYPE, index_none_values=False))
+        self.add_index('handle', multikey.UIndexDefinition(lambda obj: obj.Handle, index_none_values=False))
+        self.add_index('NODETYPE', multikey.IndexDefinition(lambda obj: obj.NODETYPE, index_none_values=False))
 
     def _save_version(self, obj: AbstractMultiStateContainer):
         self.handle_version_lookup[obj.Handle] = obj.StateVersion
@@ -214,7 +216,6 @@ class MultiStatesLookup(_MultikeyWithVersionLookup):
 
 
 class _EntityBase:
-
     def __init__(self, mdib: MdibBase, descriptor: AbstractDescriptorContainer):
         self._mdib = mdib
         self.descriptor = descriptor
@@ -233,8 +234,9 @@ class _EntityBase:
 
     def update(self):
         """Update the entity from current data in mdib."""
-        orig = self._mdib.descriptions.get_one(self.handle)
+        orig = self._mdib.descriptions.handle.get_one(self.handle)
         self.descriptor.update_from_other_container(orig)
+
 
 class Entity(_EntityBase):
     """Groups descriptor and state."""
@@ -242,7 +244,6 @@ class Entity(_EntityBase):
     def __init__(self, mdib: MdibBase, descriptor: AbstractDescriptorContainer, state: AbstractStateContainer):
         super().__init__(mdib, descriptor)
         self.state = state
-
 
     @property
     def is_multi_state(self) -> bool:
@@ -259,8 +260,12 @@ class Entity(_EntityBase):
 class MultiStateEntity(_EntityBase):
     """Groups descriptor and list of multi-states."""
 
-    def __init__(self, mdib: MdibBase, descriptor: AbstractDescriptorContainer,
-                 states: list[AbstractMultiStateContainer]):
+    def __init__(
+        self,
+        mdib: MdibBase,
+        descriptor: AbstractDescriptorContainer,
+        states: list[AbstractMultiStateContainer],
+    ):
         super().__init__(mdib, descriptor)
         self.states: dict[str, AbstractMultiStateContainer] = {s.Handle: s for s in states}
 
@@ -274,7 +279,7 @@ class MultiStateEntity(_EntityBase):
         super().update()
 
         all_orig_states = self._mdib.context_states.descriptor_handle.get(self.handle, [])
-        states_dict = { st.Handle: st for st in all_orig_states}
+        states_dict = {st.Handle: st for st in all_orig_states}
         # update existing states, remove deleted ones
         for state in list(self.states.values()):
             orig = states_dict.get(state.Handle)
@@ -283,7 +288,7 @@ class MultiStateEntity(_EntityBase):
             else:
                 self.states.pop(state.Handle)
         # add new states
-        for handle, _ in states_dict.items(): # noqa: PERF102
+        for handle, _ in states_dict.items():  # noqa: PERF102
             if handle not in self.states:
                 self.states[handle] = states_dict[handle].mk_copy()
 
@@ -305,7 +310,7 @@ class EntityGetter:
     def __init__(self, mdib: MdibBase):
         self._mdib = mdib
 
-    def by_handle(self, handle: str) ->  Entity | MultiStateEntity | None:
+    def by_handle(self, handle: str) -> Entity | MultiStateEntity | None:
         """Return entity with given handle."""
         descriptor = self._mdib.descriptions.handle.get_one(handle, allow_none=True)
         if descriptor is None:
@@ -324,7 +329,9 @@ class EntityGetter:
 
     def by_coding(self, coding: Coding) -> list[Entity | MultiStateEntity]:
         """Return all entities with descriptors type are equivalent to codeding."""
-        descriptors = [d for d in self._mdib.descriptions.objects if d.Type.is_equivalent(coding)]
+        descriptors = [
+            d for d in self._mdib.descriptions.objects if d.Type is not None and d.Type.is_equivalent(coding)
+        ]
         return [self._mk_entity(d) for d in descriptors]
 
     def by_coded_value(self, coded_value: CodedValue) -> list[Entity | MultiStateEntity]:
@@ -335,7 +342,7 @@ class EntityGetter:
     def _mk_entity(self, descriptor: AbstractDescriptorContainer) -> Entity | MultiStateEntity:
         if descriptor.is_context_descriptor:
             states = self._mdib.context_states.descriptor_handle.get(descriptor.Handle, [])
-            return MultiStateEntity(self._mdib,copy.deepcopy(descriptor), copy.deepcopy(states))
+            return MultiStateEntity(self._mdib, copy.deepcopy(descriptor), copy.deepcopy(states))
         state = self._mdib.states.descriptor_handle.get_one(descriptor.Handle)
         return Entity(self._mdib, copy.deepcopy(descriptor), copy.deepcopy(state))
 
@@ -347,7 +354,6 @@ class EntityGetter:
     def __len__(self) -> int:
         """Return number of entities."""
         return len(self._mdib.descriptions.objects)
-
 
 
 class MdibBase:
@@ -365,7 +371,8 @@ class MdibBase:
     updated_descriptors_by_handle = properties.ObservableProperty(fire_only_on_changed_value=False)
     deleted_descriptors_by_handle = properties.ObservableProperty(fire_only_on_changed_value=False)
     deleted_states_by_handle = properties.ObservableProperty(
-        fire_only_on_changed_value=False)  # is a result of deleted descriptors
+        fire_only_on_changed_value=False,
+    )  # is a result of deleted descriptors
     description_modifications = properties.ObservableProperty(fire_only_on_changed_value=False)
     operation_by_handle = properties.ObservableProperty(fire_only_on_changed_value=False)
     sequence_id = properties.ObservableProperty()
@@ -391,7 +398,6 @@ class MdibBase:
         self.mdstate_version = 0
         self.mddescription_version = 0
 
-
     @property
     def logger(self) -> LoggerAdapter:
         """Return the logger."""
@@ -399,7 +405,7 @@ class MdibBase:
 
     @property
     def mdib_version_group(self) -> MdibVersionGroup:
-        """"Get current version data."""
+        """Get current version data."""
         return MdibVersionGroup(self.mdib_version, self.sequence_id, self.instance_id)
 
     def add_description_containers(self, description_containers: list[AbstractDescriptorContainer]):
@@ -433,19 +439,23 @@ class MdibBase:
 
     def _set_descriptor_container_reference(self, state_container: AbstractStateContainer):
         """Set state_container.descriptor_container if all is fine, otherwise logs error."""
-        descriptor_container = self.descriptions.handle.get_one(state_container.DescriptorHandle,
-                                                                allow_none=True)
+        descriptor_container = self.descriptions.handle.get_one(state_container.DescriptorHandle, allow_none=True)
         if descriptor_container is None:
             self._logger.warning(  # noqa: PLE1205
                 'state "{}" (type={}) has no descriptor in mdib!',
-                state_container.DescriptorHandle, state_container.NODETYPE)
+                state_container.DescriptorHandle,
+                state_container.NODETYPE,
+            )
         elif descriptor_container.DescriptorVersion == state_container.DescriptorVersion:
             state_container.descriptor_container = descriptor_container
         else:
             self._logger.warning(  # noqa: PLE1205
                 'state "{}" (type={}) : descriptor version expect "{}", found "{}"',
-                state_container.DescriptorHandle, state_container.NODETYPE,
-                descriptor_container.DescriptorVersion, state_container.DescriptorVersion)
+                state_container.DescriptorHandle,
+                state_container.NODETYPE,
+                descriptor_container.DescriptorVersion,
+                state_container.DescriptorVersion,
+            )
 
     def add_state_containers(self, state_containers: list[AbstractStateContainer | AbstractMultiStateContainer]):
         """Add states to self.states and self.context_states.
@@ -461,29 +471,41 @@ class MdibBase:
                 my_multikey.add_object(state_container)
             except KeyError as ex:
                 if state_container.is_context_state:
-                    self._logger.error('add_state_containers: {}, Handle={}; {}',  # noqa: PLE1205 TRY400
-                                       ex, state_container.Handle, traceback.format_exc())
+                    self._logger.error(  # noqa: PLE1205 TRY400
+                        'add_state_containers: {}, Handle={}; {}',
+                        ex,
+                        state_container.Handle,
+                        traceback.format_exc(),
+                    )
                 else:
-                    self._logger.error('add_state_containers: {}, DescriptorHandle={}; {}',  # noqa: PLE1205 TRY400
-                                       ex, state_container.DescriptorHandle, traceback.format_exc())
+                    self._logger.error(  # noqa: PLE1205 TRY400
+                        'add_state_containers: {}, DescriptorHandle={}; {}',
+                        ex,
+                        state_container.DescriptorHandle,
+                        traceback.format_exc(),
+                    )
 
     def _reconstruct_md_description(self) -> xml_utils.LxmlElement:
         """Build dom tree of descriptors from current data."""
         pm = self.data_model.pm_names
         doc_nsmap = self.nsmapper.ns_map
         root_containers = self.descriptions.parent_handle.get(None) or []
-        md_description_node = etree.Element(pm.MdDescription,
-                                            attrib={'DescriptionVersion': str(self.mddescription_version)},
-                                            nsmap=doc_nsmap)
+        md_description_node = etree.Element(
+            pm.MdDescription,
+            attrib={'DescriptionVersion': str(self.mddescription_version)},
+            nsmap=doc_nsmap,
+        )
         for root_container in root_containers:
             self.make_descriptor_node(root_container, md_description_node, tag=pm.Mds, set_xsi_type=False)
         return md_description_node
 
-    def make_descriptor_node(self,
-                             descriptor_container: AbstractDescriptorContainer,
-                             parent_node: xml_utils.LxmlElement,
-                             tag: etree.QName,
-                             set_xsi_type: bool = True) -> xml_utils.LxmlElement:
+    def make_descriptor_node(
+        self,
+        descriptor_container: AbstractDescriptorContainer,
+        parent_node: xml_utils.LxmlElement,
+        tag: etree.QName,
+        set_xsi_type: bool = True,
+    ) -> xml_utils.LxmlElement:
         """Create a lxml etree node with subtree from instance data.
 
         :param descriptor_container: a descriptor container instance
@@ -492,12 +514,12 @@ class MdibBase:
         :param set_xsi_type: if true, the NODETYPE will be used to set the xsi:type attribute of the node
         :return: an etree node.
         """
-        ns_map = self.nsmapper.partial_map(self.nsmapper.PM, self.nsmapper.XSI) \
-            if set_xsi_type else self.nsmapper.partial_map(self.nsmapper.PM)
-        node = etree.SubElement(parent_node,
-                                tag,
-                                attrib={'Handle': descriptor_container.Handle},
-                                nsmap=ns_map)
+        ns_map = (
+            self.nsmapper.partial_map(self.nsmapper.PM, self.nsmapper.XSI)
+            if set_xsi_type
+            else self.nsmapper.partial_map(self.nsmapper.PM)
+        )
+        node = etree.SubElement(parent_node, tag, attrib={'Handle': descriptor_container.Handle}, nsmap=ns_map)
         descriptor_container.update_node(node, self.nsmapper, set_xsi_type)  # create all
         child_list = self.descriptions.parent_handle.get(descriptor_container.Handle, [])
         # append all child containers, then bring all child elements in correct order
@@ -524,9 +546,12 @@ class MdibBase:
         mdib_node.append(md_description_node)
 
         # add a list of states
-        md_state_node = etree.SubElement(mdib_node, pm.MdState,
-                                         attrib={'StateVersion': str(self.mdstate_version)},
-                                         nsmap=doc_nsmap)
+        md_state_node = etree.SubElement(
+            mdib_node,
+            pm.MdState,
+            attrib={'StateVersion': str(self.mdstate_version)},
+            nsmap=doc_nsmap,
+        )
         tag = pm.State
         for state_container in self.states.objects:
             md_state_node.append(state_container.mk_state_node(tag, self.nsmapper))
@@ -564,10 +589,12 @@ class MdibBase:
         with_types = [d for d in descriptors if d.Type is not None]
         return [d for d in with_types if have_matching_codes(d.Type, code)]
 
-    def get_metric_descriptor_by_code(self,
-                                      vmd_code: [Coding, CodedValue],
-                                      channel_code: [Coding, CodedValue],
-                                      metric_code: [Coding, CodedValue]) -> AbstractDescriptorContainer | None:
+    def get_metric_descriptor_by_code(
+        self,
+        vmd_code: [Coding, CodedValue],
+        channel_code: [Coding, CodedValue],
+        metric_code: [Coding, CodedValue],
+    ) -> AbstractDescriptorContainer | None:
         """get_metric_descriptor_by_code is the "correct" way to find a descriptor.
 
         Using handles is shaky, because they have no meaning and can change over time!
@@ -586,10 +613,12 @@ class MdibBase:
                     raise ValueError(msg)
         return None
 
-    def get_operations_for_metric(self,
-                                  vmd_code: [Coding, CodedValue],
-                                  channel_code: [Coding, CodedValue],
-                                  metric_code: [Coding, CodedValue]) -> list[AbstractDescriptorContainer]:
+    def get_operations_for_metric(
+        self,
+        vmd_code: [Coding, CodedValue],
+        channel_code: [Coding, CodedValue],
+        metric_code: [Coding, CodedValue],
+    ) -> list[AbstractDescriptorContainer]:
         """get_operations_for_metric is the "correct" way to find an operation.
 
         Using handles is shaky, because they have no meaning and can change over time!
@@ -597,8 +626,11 @@ class MdibBase:
         descriptor_container = self.get_metric_descriptor_by_code(vmd_code, channel_code, metric_code)
         return self.get_operation_descriptors_for_descriptor_handle(descriptor_container.Handle)
 
-    def get_operation_descriptors_for_descriptor_handle(self, descriptor_handle: str,
-                                                        **additional_filters: Any) -> list[AbstractDescriptorContainer]:
+    def get_operation_descriptors_for_descriptor_handle(
+        self,
+        descriptor_handle: str,
+        **additional_filters: Any,
+    ) -> list[AbstractDescriptorContainer]:
         """Get operation descriptors that have descriptor_handle as OperationTarget.
 
         :param descriptor_handle: the handle for that operations shall be found
@@ -616,13 +648,15 @@ class MdibBase:
         """Get a list of all operation descriptors."""
         pm = self.data_model.pm_names
         result = []
-        for node_type in (pm.SetValueOperationDescriptor,
-                          pm.SetStringOperationDescriptor,
-                          pm.ActivateOperationDescriptor,
-                          pm.SetContextStateOperationDescriptor,
-                          pm.SetMetricStateOperationDescriptor,
-                          pm.SetComponentStateOperationDescriptor,
-                          pm.SetAlertStateOperationDescriptor):
+        for node_type in (
+            pm.SetValueOperationDescriptor,
+            pm.SetStringOperationDescriptor,
+            pm.ActivateOperationDescriptor,
+            pm.SetContextStateOperationDescriptor,
+            pm.SetMetricStateOperationDescriptor,
+            pm.SetComponentStateOperationDescriptor,
+            pm.SetAlertStateOperationDescriptor,
+        ):
             result.extend(self.descriptions.NODETYPE.get(node_type, []))
         return result
 
@@ -653,13 +687,17 @@ class MdibBase:
                 for handle in all_handles:
                     selected_objects.extend(self.descriptions.parent_handle.get(handle, []))
             # filter current list
-            selected_objects = [o for o in selected_objects if
-                                o.Type is not None and have_matching_codes(o.Type, coding)]
+            selected_objects = [
+                o for o in selected_objects if o.Type is not None and have_matching_codes(o.Type, coding)
+            ]
         return selected_objects
 
-    def get_all_descriptors_in_subtree(self, root_descriptor_container: AbstractDescriptorContainer,
-                                       depth_first: bool = True,
-                                       include_root: bool = True) -> list[AbstractDescriptorContainer]:
+    def get_all_descriptors_in_subtree(
+        self,
+        root_descriptor_container: AbstractDescriptorContainer,
+        depth_first: bool = True,
+        include_root: bool = True,
+    ) -> list[AbstractDescriptorContainer]:
         """Return the tree below descriptor_container as a flat list.
 
         :param root_descriptor_container: root descriptor
@@ -690,8 +728,11 @@ class MdibBase:
         deleted_descriptors = {}
         deleted_states = {}
         for descriptor_container in descriptor_containers:
-            self._logger.debug('rm Descriptor node {} handle {}',  # noqa: PLE1205
-                               descriptor_container.NODETYPE, descriptor_container.Handle)
+            self._logger.debug(  # noqa: PLE1205
+                'rm Descriptor node {} handle {}',
+                descriptor_container.NODETYPE,
+                descriptor_container.Handle,
+            )
             self.descriptions.remove_object(descriptor_container)
             deleted_descriptors[descriptor_container.Handle] = descriptor_container
             for m_key in (self.states, self.context_states):
@@ -699,8 +740,11 @@ class MdibBase:
                 if state_containers is not None:
                     # make a copy, otherwise remove_objects will manipulate same list in place
                     state_containers = state_containers[:]
-                    self._logger.debug('rm {} states(s) associated to descriptor {} ',  # noqa: PLE1205
-                                       len(state_containers), descriptor_container.Handle)
+                    self._logger.debug(  # noqa: PLE1205
+                        'rm {} states(s) associated to descriptor {} ',
+                        len(state_containers),
+                        descriptor_container.Handle,
+                    )
                     m_key.remove_objects(state_containers)
                     deleted_states[descriptor_container.Handle] = state_containers