+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - renamed parameter in ``SdcConsumer.do_subscribe`` from `expire_minutes` to `expire_seconds`. It was already handled as seconds but was named wrong [#436](https://github.com/Draegerwerk/sdc11073/pull/436)
+- increase the default timeout for starting the HTTP server and make this timeout configurable to mitigate startup delays, issue [#320](https://github.com/Draegerwerk/sdc11073/issues/320)
 
 ### Fixed
 
@@ -28,6 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `isoduration.UTC` class. Use `datetime.timezone.utc` instead. [#435](https://github.com/Draegerwerk/sdc11073/pull/445)
 - support for python 3.9
+- MDIBs with entity handling [#462](https://github.com/Draegerwerk/sdc11073/pull/462)
 
 ## [v2.3.0] - 2025-07-08